+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - `finch` `Mxpanel.HTTPClient` adapter.
+- `Mxpanel.track_many/2`.
 
 ## [0.1.0] - 2021-06-16